Debug fine-grained RBAC
